+++ plugin/src/main/java/io/snyk/eclipse/plugin/views/snyktoolview/filters/FixableFilter.java
@@ -1,28 +1,30 @@
 package io.snyk.eclipse.plugin.views.snyktoolview.filters;
 
-import io.snyk.eclipse.plugin.preferences.Preferences;
+import static io.snyk.eclipse.plugin.preferences.Preferences.FILTER_SHOW_ONLY_FIXABLE;
+
+import java.util.function.Predicate;
+
 import io.snyk.eclipse.plugin.views.snyktoolview.TreeFilterManager;
+import io.snyk.languageserver.protocolextension.messageObjects.scanResults.Issue;
 
-public class FixableFilter implements BaseFilter {
-	private TreeFilterManager filterManager;
-	private Preferences preferences;
-	private String preferenceKey;
-
-	public FixableFilter(TreeFilterManager filterManager, Preferences preferences, String preferenceKey) {
-		this.filterManager = filterManager;
-		this.preferences = preferences;
-		this.preferenceKey = preferenceKey;
-
+public class FixableFilter extends BaseFilter {	
+	private static final Predicate<Issue> predicate = issue -> issue.hasFix();
+	
+	public FixableFilter(TreeFilterManager tfm) {
+		super(FILTER_SHOW_ONLY_FIXABLE, predicate, tfm);
 	}
 
 	@Override
 	public void applyFilter() {
-		boolean booleanPref = this.preferences.getBooleanPref(this.preferenceKey);
+		// this is inverse to the other filters, so we need to overwrite the BaseFilter logic
+		boolean booleanPref = preferences.getBooleanPref(this.filterName);
 
 		if (booleanPref) {
-			this.filterManager.addTreeFilter(this.preferenceKey, issue -> issue.hasFix());
+			this.filterManager.addTreeFilter(this.filterName, predicate);
 		} else {
-			this.filterManager.removeTreeFilter(this.preferenceKey);
+			this.filterManager.removeTreeFilter(this.filterName);
 		}
 	}
+	
+	
 }