+++ plugin/src/main/java/io/snyk/eclipse/plugin/views/snyktoolview/filters/IgnoresIgnoredIssuesFilter.java
@@ -7,10 +7,11 @@ import java.util.function.Predicate;
 import io.snyk.eclipse.plugin.views.snyktoolview.TreeFilterManager;
 import io.snyk.languageserver.protocolextension.messageObjects.scanResults.Issue;
 
-public class IgnoresOpenIssuesFilter extends BaseFilter {	
-	private static final Predicate<Issue> predicate = issue -> !issue.isIgnored();
-	
-	public IgnoresOpenIssuesFilter(TreeFilterManager tfm) {
+public class IgnoresIgnoredIssuesFilter extends BaseFilter {
+	private static final Predicate<Issue> predicate = issue -> issue.isIgnored();
+
+	public IgnoresIgnoredIssuesFilter(TreeFilterManager tfm) {
 		super(FILTER_IGNORES_SHOW_IGNORED_ISSUES, predicate, tfm);
 	}
+
 }