+++ plugin/src/main/java/io/snyk/eclipse/plugin/views/snyktoolview/handlers/EnableCodeQualityProductHandler.java
@@ -1,18 +1,13 @@
 package io.snyk.eclipse.plugin.views.snyktoolview.handlers;
 
+import static io.snyk.eclipse.plugin.preferences.Preferences.ACTIVATE_SNYK_CODE_QUALITY;
+
 import org.eclipse.ui.commands.IElementUpdater;
 
-import io.snyk.eclipse.plugin.preferences.Preferences;
-import io.snyk.eclipse.plugin.utils.SnykIcons;
-
-public class EnableCodeQualityProductHandler extends BaseHandler implements IElementUpdater {
+public class EnableCodeQualityProductHandler extends BaseProductFilterHandler implements IElementUpdater {
 
 	public EnableCodeQualityProductHandler() {
-		super();
-
-		iconEnabled = SnykIcons.ENABLED;
-		iconDisabled = SnykIcons.DISABLED;
-		preferenceKey = Preferences.ACTIVATE_SNYK_CODE_QUALITY;
+		super(ACTIVATE_SNYK_CODE_QUALITY);
+		preferenceKey = ACTIVATE_SNYK_CODE_QUALITY;
 	}
-
 }