+++ tests/src/test/java/io/snyk/languageserver/LsConfigurationUpdaterTest.java
@@ -114,7 +114,7 @@ class LsConfigurationUpdaterTest {
 		when(preferenceMock.getPref(Preferences.AUTH_TOKEN_KEY, "")).thenReturn("my-token");
 		when(preferenceMock.getBooleanPref(Preferences.SCANNING_MODE_AUTOMATIC)).thenReturn(true);
 		when(preferenceMock.getBooleanPref(Preferences.USE_TOKEN_AUTH, false)).thenReturn(true);
-		when(preferenceMock.getPref(Preferences.FILTER_DELTA_NEW_ISSUES, Boolean.FALSE.toString())).thenReturn("true");
+		when(preferenceMock.getPref(Preferences.ENABLE_DELTA, Boolean.FALSE.toString())).thenReturn("true");
 
 		FolderConfigsParam mockFolderConfigsParam = new FolderConfigsParam(new ArrayList<>());
 		when(mockFolderConfigs.updateFolderConfigs()).thenReturn(mockFolderConfigsParam);